+++ optim/vars.go
@@ -373,6 +373,6 @@ Description:
 
 	Returns the dimension of the Variable object (should be scalar).
 */
-func (v Variable) Dims() []uint64 {
-	return []uint64{1, 1}
+func (v Variable) Dims() []uint {
+	return []uint{1, 1}
 }
